Refactor getDisplayName to its own file
